refactor(#120): move UCSC Cell Browser constants to skills/ucsc_cellbrowser/shared/constants.py
Plugin-local UCSC config (base URL, file names, JSON path, column names)
belongs to the ucsc_cellbrowser plugin, not core. Re-expressed
UCSC_JSON_FILE_PATH using a parents[3] resolution from the new location
(same final target: hvantk/resources/cells_ucsc_datasets.json). Dropped
the module-import debug-log dumps. Updated 6 Python consumers (cli.py,
3 test modules, shared/datasets.py, shared/ucsc.py) and the SKILL.md
path reference. One of several PRs closing #120.
