+++ hvantk/core/constants.py
@@ -9,22 +9,6 @@ logger = logging.getLogger(__name__)
 BASE_DIR = Path(__file__).resolve().parent
 logger.debug(f"Base directory: {BASE_DIR}")
 
-## UCSC Cell Browser base URL
-UCSC_CELL_BROWSER_BASE_URL = "https://cells.ucsc.edu"
-logger.debug(f"UCSC Cell Browser base URL: {UCSC_CELL_BROWSER_BASE_URL}")
-EXPRESSION_MATRIX_FILE_NAME = "exprMatrix.tsv.gz"
-logger.debug(f"Expression matrix file name: {EXPRESSION_MATRIX_FILE_NAME}")
-METADATA_FILE_NAME = "meta.tsv"
-logger.debug(f"Metadata file name: {METADATA_FILE_NAME}")
-
-# Path to the JSON file containing the UCSC cell datasets
-UCSC_JSON_FILE_PATH = BASE_DIR.parent / "resources" / "cells_ucsc_datasets.json"
-logger.debug(f"UCSC JSON file path: {UCSC_JSON_FILE_PATH}")
-
-# UCSC gene and cell ID columns
-UCSC_CELL_ID_COLUMN = "cell_id"
-UCSC_GENE_COLUMN = "gene"
-
 ## Expression Atlas base URL
 EXPRESSION_ATLAS_BASE_URL = "https://www.ebi.ac.uk/gxa/experiments-content"
 logger.debug(f"Expression Atlas base URL: {EXPRESSION_ATLAS_BASE_URL}")
@@ -234,12 +218,6 @@ ALPHAGENOME_DEFAULT_REQUEST_TIMEOUT = 120
 # Explicit public API for this module
 __all__ = [
     "BASE_DIR",
-    "UCSC_CELL_BROWSER_BASE_URL",
-    "EXPRESSION_MATRIX_FILE_NAME",
-    "METADATA_FILE_NAME",
-    "UCSC_JSON_FILE_PATH",
-    "UCSC_CELL_ID_COLUMN",
-    "UCSC_GENE_COLUMN",
     "EXPRESSION_ATLAS_BASE_URL",
     "REGISTRY_ROOT_PATH",
     "TRANSCRIPTOMICS_DATASETS_PATH",
