+++ hvantk/skills/ucsc_cellbrowser/SKILL.md
@@ -60,7 +60,7 @@ When invoked to build or update:
 
 1. Verify `anndata`, `scanpy`, and `scipy` are importable. The current SessionStart hook (`hvantk/skills/_hooks/check-hail.sh`) only checks Hail + Java — anndata/scanpy availability is a known gap; verify manually with `python -c "import anndata, scanpy, scipy"` until the hook is extended.
 2. Confirm the expression input is a UCSC matrix: read the first line; the leading column should be a gene identifier (e.g., `gene`/`geneId`/`Symbol`), not a cell ID. The remaining header fields should be cell IDs that match the metadata TSV's cell-ID column.
-3. Confirm the metadata input: read the first line of the metadata TSV; the cell-ID column is whatever `UCSC_CELL_ID_COLUMN` resolves to (see `hvantk/core/constants.py`); other columns become `obs` fields.
+3. Confirm the metadata input: read the first line of the metadata TSV; the cell-ID column is whatever `UCSC_CELL_ID_COLUMN` resolves to (see `hvantk/skills/ucsc_cellbrowser/shared/constants.py`); other columns become `obs` fields.
 4. Call `build_ucsc_ad(expression_matrix_path=..., metadata_path=..., output_path=..., gene_column=..., split_gene_field=..., overwrite=..., backed=None, column_batch=64)`. The function auto-selects backed mode based on file size; pass `backed=True` to force.
 5. If backed mode is selected (auto or forced), `output_path` is required. Division of responsibility: `build_ucsc_atlas_backed` writes CSC columns directly to disk (no return value of interest); `build_ucsc_ad` re-opens the written file via `anndata.read_h5ad(output_path, backed='r')` and returns that read-backed handle. Per § 4, do not rely on `annotate_column_summary_ad`-derived fields in this branch — the backed path skips that pass.
 6. Backed-path operation order is fixed: coerce `obs` via `coerce_obs_for_h5ad` → pre-flight via `_probe_write_elem` (catches obs-serialization failures BEFORE streaming any X columns) → stream X column-batches via `write_elem` + `sparse_dataset(...).append(...)` → finalize by writing `/obs`, `/var`, `/uns`. Reordering risks committing a partial atlas to disk before discovering an obs failure.
