+++ hvantk/skills/ucsc_cellbrowser/cli.py
@@ -16,7 +16,7 @@ logger = logging.getLogger(__name__)
 from hvantk.skills.ucsc_cellbrowser.shared.datasets import UCSCDataSetCollection
 from hvantk.core.config import CONTEXT_SETTINGS
 import hvantk.core.utils.file_utils as file_utils
-from hvantk.core.constants import (
+from hvantk.skills.ucsc_cellbrowser.shared.constants import (
     UCSC_CELL_BROWSER_BASE_URL,
     EXPRESSION_MATRIX_FILE_NAME,
     METADATA_FILE_NAME,
