+++ hvantk/skills/ucsc_cellbrowser/tests/test_downloader.py
@@ -6,7 +6,7 @@ from unittest.mock import patch, MagicMock
 from click.testing import CliRunner
 from hvantk.core.utils.file_utils import download_file
 from hvantk.skills.ucsc_cellbrowser.cli import ucsc_downloader
-from hvantk.core.constants import (
+from hvantk.skills.ucsc_cellbrowser.shared.constants import (
     UCSC_CELL_BROWSER_BASE_URL,
     EXPRESSION_MATRIX_FILE_NAME,
     METADATA_FILE_NAME,
