+++ hvantk/tests/test_cli_downloader.py
@@ -4,7 +4,7 @@ from unittest.mock import patch, MagicMock
 import pytest
 from click.testing import CliRunner
 from hvantk.skills.ucsc_cellbrowser.cli import ucsc_downloader
-from hvantk.core.constants import UCSC_CELL_BROWSER_BASE_URL
+from hvantk.skills.ucsc_cellbrowser.shared.constants import UCSC_CELL_BROWSER_BASE_URL
 
 
 @pytest.fixture
